MOVE SideDrawer to components
