+++ src/components/SideDrawer/sidedrawer.css
@@ -1,0 +1,16 @@
+.MuiListItemText-root a {
+	color: white;
+}
+
+.MuiPaper-root {
+	background-color: #292b2f !important;
+}
+
+.MuiListItemIcon-root {
+	color: white !important;
+	font-size: 23px !important;
+}
+
+.MuiSvgIcon-root {
+	color: white;
+}
